+++ packages/destination-actions/src/destinations/google-enhanced-conversions/generated-types.ts
@@ -5,10 +5,8 @@ export interface Settings {
    * You will find this information in the event snippet for your conversion action, for example `send_to: AW-CONVERSION_ID/AW-CONVERSION_LABEL`. In the sample snippet, AW-CONVERSION_ID stands for the conversion ID unique to your account. Enter the conversion Id, without the AW- prefix. **Required if you are using an Action that sends data to the legacy Google Enhanced Conversions API.**
    */
   conversionTrackingId?: string
-
   /**
    * ID of your Google Ads Account. This should be 10-digits and in XXX-XXX-XXXX format. **Required if you are using an Action that sends data to the Google Ads API.**
    */
   customerId?: string
-
 }
